refactor: move files and change insight class
