+++ ee/clickhouse/queries/test/test_retention.py
@@ -1,5 +1,5 @@
 from ee.clickhouse.models.group import create_group
-from ee.clickhouse.queries.retention.clickhouse_retention import ClickhouseRetention
+from ee.clickhouse.queries.retention import ClickhouseRetention
 from ee.clickhouse.util import ClickhouseTestMixin, snapshot_clickhouse_queries
 from posthog.models.action import Action
 from posthog.models.action_step import ActionStep
