+++ posthog/queries/retention/retention.py
@@ -1,18 +1,23 @@
 from typing import Any, Dict, List, NamedTuple, Optional, Tuple, Union
 from urllib.parse import urlencode
 
-from ee.clickhouse.queries.retention.retention_event_query import RetentionEventsQuery
 from ee.clickhouse.sql.retention.retention import RETENTION_BREAKDOWN_SQL
 from posthog.client import substitute_params, sync_execute
 from posthog.constants import RETENTION_FIRST_TIME, RetentionQueryType
 from posthog.models.filters.retention_filter import RetentionFilter
 from posthog.models.team import Team
+from posthog.queries.retention.actors_query import RetentionActors, RetentionActorsByPeriod, build_actor_activity_query
+from posthog.queries.retention.event_query import RetentionEventsQuery
 
 BreakdownValues = Tuple[Union[str, int], ...]
 CohortKey = NamedTuple("CohortKey", (("breakdown_values", BreakdownValues), ("period", int)))
 
 
-class ClickhouseRetention:
+class Retention:
+    event_query = RetentionEventsQuery
+    actors_query = RetentionActors
+    actors_by_period_query = RetentionActorsByPeriod
+
     def __init__(self, base_uri="/"):
         self._base_uri = base_uri
 
@@ -26,7 +31,6 @@ class ClickhouseRetention:
     def _get_retention_by_breakdown_values(
         self, filter: RetentionFilter, team: Team,
     ) -> Dict[CohortKey, Dict[str, Any]]:
-        from ee.clickhouse.queries.retention.retention_actors import build_actor_activity_query
 
         actor_query = build_actor_activity_query(filter=filter, team=team)
 
@@ -111,9 +115,8 @@ class ClickhouseRetention:
         return result
 
     def actors(self, filter: RetentionFilter, team: Team):
-        from ee.clickhouse.queries.retention.retention_actors import ClickhouseRetentionActors
 
-        _, serialized_actors = ClickhouseRetentionActors(team=team, filter=filter).get_actors()
+        _, serialized_actors = self.actors_query(team=team, filter=filter).get_actors()
 
         return serialized_actors
 
@@ -136,9 +139,7 @@ class ClickhouseRetention:
         interval, where the index of the list is the interval it refers to.
         """
 
-        from ee.clickhouse.queries.retention.retention_actors import ClickhouseRetentionActorsByPeriod
-
-        return ClickhouseRetentionActorsByPeriod(team=team, filter=filter).actors()
+        return self.actors_by_period_query(team=team, filter=filter).actors()
 
 
 def build_returning_event_query(
